Fix bug due to interaction of partition table & string interning
A user reported that the previous commit (03fca656d70) — which
added a partition table for USB drive booting — was causing
SectorLISP to crash on the input
	(CONS (QUOTE AAA) NIL)

This turned out to be because the SectorLISP code interns new
tokens starting at an area where it sees two null bytes.  The
newly added "partition table" entry for a bootable partition
contains multiple 00's in a row, & was followed by code.  The
code was clobbered.

This commit moves the bootable partition table entry towards
the end of the partition table, which is safe to clobber.